fix(ProjectUser): route nested

+++ src/main/java/com/monza96/backend/services/ProjectUserService.java
@@ -27,26 +27,40 @@ public class ProjectUserService {
     private final RoleService roleService;
     private final UserService userService;
 
-    public List<ProjectUserResponseDTO> findAll() {
-        return projectUserRepository.findAll().stream().map(x -> ProjectUserMapper.toResponseDTO(x)).toList();
+    public List<ProjectUserResponseDTO> findAll(Long projectId) {
+        return projectUserRepository.findAllByProjectId(projectId)
+                .stream()
+                .map(x -> ProjectUserMapper.toResponseDTO(x)).toList();
     }
 
-    public ProjectUserResponseDTO findById(Long id) {
-        ProjectUser projectUser = findEntityById(id);
+    public ProjectUserResponseDTO findById(Long projectId, Long id) {
+        ProjectUser projectUser = findEntityByProjectIdAndId(projectId, id);
         return ProjectUserMapper.toResponseDTO(projectUser);
     }
 
-    public ProjectUserResponseDTO create(ProjectUserRequestDTO dto) {
-        //every user should send all the data
+    public ProjectUserResponseDTO create(Long projectId, ProjectUserRequestDTO dto) {
+        Project project = projectService.findEntityById(projectId);
+
         User user = userService.findEntityById(dto.userId());
+        project.getProjectUsers().stream()
+                .filter(x -> x.getUser().equals(user)).findFirst()
+                .ifPresent(x -> {
+                    throw new DatabaseException("User already in project, use PUT to update the role");
+                });
+
         Role role = roleService.findEntityById(dto.roleId());
-        Project project = projectService.findEntityById(dto.projectId());
 
         ProjectUser projectUser = new ProjectUser(null, user, project, role);
+
         return ProjectUserMapper.toResponseDTO(projectUserRepository.save(projectUser));
     }
 
-    public void delete(Long id) {
+    public void delete(Long projectId, Long id) {
+        ProjectUser projectUser = findEntityById(id);
+        if (!projectUser.getProject().getId().equals(projectId)) {
+            throw new ResourceNotFoundException(ProjectUser.class, id);
+        }
+
         try {
             projectUserRepository.deleteById(id);
         } catch (EmptyResultDataAccessException e) {
@@ -56,11 +70,14 @@ public class ProjectUserService {
         }
     }
 
-    public ProjectUserResponseDTO update(Long id, ProjectUserRequestDTO dto) {
-        ProjectUser projectUser = findEntityById(id);
+    public ProjectUserResponseDTO update(Long projectId, Long id, ProjectUserRequestDTO dto) {
+        ProjectUser projectUser = findEntityByProjectIdAndId(projectId, id);
 
-        //TODO add more logic here, an user may not self update the role
-        //can only update the role
+        User userUpdater = userService.findEntityById(dto.userId());
+        if (projectUser.getUser().equals(userUpdater)) {
+            throw new DatabaseException("An user may not update its own role in a project");
+        }
+
         projectUser.setRole(roleService.findEntityById(dto.roleId()));
 
         return ProjectUserMapper.toResponseDTO(projectUserRepository.save(projectUser));
@@ -68,6 +85,11 @@ public class ProjectUserService {
 
     ProjectUser findEntityById(Long id) {
         return projectUserRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException(ProjectUser.class, id));
+    }
+
+    ProjectUser findEntityByProjectIdAndId(Long id, Long projectId) {
+        return projectUserRepository.findByProjectIdAndId(projectId, id)
                 .orElseThrow(() -> new ResourceNotFoundException(ProjectUser.class, id));
     }
 }
